pyproject.toml
[project]
name = "matrixamnesia"
version = "0.1.0"
description = "MatrixAmnesia is a simple script for removing user messages from Matrix room"
requires-python = ">=3.14"
dependencies = [
    "matrix-nio>=0.25.2",
]
